+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changes
 
 ## Unreleased ~ master
+
+## 1.2.0
  - Add a collection macro for `BinaryHeap`: `bheap!`
  - Add a collection macro for `VecDeque`: `deque!`
  - Add empty pattern for collections macros already created